Created scenes, prefabs, scripts, and materials
Created all scenes for game overs and completions, as well as working buttons to transition between scenes. Created a functional moving camera for players to view levels and clickable areas for players to set platforms. Added a door for the goal, and a spring prefab for the players to use in their platforms.